+++ CHANGELOG.md
@@ -1,2 +1,7 @@
-# 16.12.2023 11:25
+# 17.12.2023 4:12AM
+- [feat]: Doctor getPatientVitals implemented
+- [feat]: checkin, checkout implemented for patient
+- [feat]: activePatientEId implemented
+
+# 16.12.2023 11:25AM
 - [feat]: DioExecutor created & basic scaffolding complete